Converting to tensorrt

+++ detectron2/export/torchscript.py
@@ -54,7 +54,7 @@ def scripting_with_instances(model, fields):
     ), "Currently we only support exporting models in evaluation mode to torchscript"
 
     with freeze_training_mode(model), patch_instances(fields):
-        scripted_model = torch.jit.script(model)
+        scripted_model = torch.jit.script(model.float())
         return scripted_model
